epic refactor
